break up util functions into smaller conceptual modules
