+++ src/common/util.ts
@@ -1,57 +1,8 @@
-import type * as t from '../types';
-import { R, crypto, fs, fsPath } from './libs';
+import { fs } from './libs';
 
-export const isNothing = (value: any) => R.isNil(value) || R.isEmpty(value);
-export const isString = R.is(String);
-
-export const toAbsolutePath = (path: string) => {
-  return path.startsWith('.') ? fsPath.resolve(path) : path;
-};
-
-export const ensureString = (defaultValue: string, text?: string): string => {
-  return typeof text === 'string' ? text : defaultValue;
-};
-
-export const compact = (input: any[]): string[] => {
-  const flat = [].concat(...input);
-  return flat.filter((value) => !R.isNil(value));
-};
-
-export const toStringArray = R.pipe(compact, R.map(R.toString));
-
-export const isFileSync = (path: string) => {
-  return fs.existsSync(path) ? fs.lstatSync(path).isFile() : false;
-};
-
-export const readFileSync = (path: string) => {
-  return fs.existsSync(path) ? fs.readFileSync(path).toString() : undefined;
-};
-
-export const filePathsP = async (basePath: string, ns: string): Promise<string[]> => {
-  if (!(await fs.pathExists(basePath))) return [];
-  return (await fs.readdir(basePath))
-    .filter(Boolean)
-    .filter((name) => (ns ? name.startsWith(ns) : true))
-    .filter((name) => (!ns ? !name.includes('-') : true))
-    .map((name) => `${basePath}/${name}`);
-};
-
-/**
- * Turns a set of values into a HEX hash code.
- * @param values: The set of values to hash.
- */
-export const hash = (algorithm: t.HashAlgorithm, ...values: any[]) => {
-  if (R.pipe(compact, R.isEmpty)(values)) return undefined;
-  const resultHash = crypto.createHash(algorithm);
-  const addValue = (value: any) => resultHash.update(value);
-  const addValues = R.forEach(addValue);
-  R.pipe(toStringArray, addValues)(values);
-  return resultHash.digest('hex');
-};
-
-export const hashExists = (algorithm: t.HashAlgorithm) => {
-  return crypto.getHashes().includes(algorithm);
-};
+export * from './util.crypto';
+export * from './util.fs';
+export * from './util.primitive';
 
 /**
  * Retrieve a value from the given path.
@@ -74,22 +25,16 @@ export async function getValueP(path: string, defaultValue?: any) {
 /**
  * Format value structure.
  */
-export const toGetValue = (data: any) => {
+export function toGetValue(data: any) {
   if (isExpired(data)) throw new Error(`Cache item has expired.`);
   if (data.type === 'Date') return new Date(data.value);
   return data.value;
-};
-
-/**
- * Stringify a value into JSON.
- */
-export const toJson = (value: any, ttl: number) =>
-  JSON.stringify({ value, type: R.type(value), created: new Date(), ttl });
+}
 
 /**
  * Check's a cache item to see if it has expired.
  */
-export const isExpired = (data: any) => {
+export function isExpired(data: any) {
   const timeElapsed = (new Date().getTime() - new Date(data.created).getTime()) / 1000;
   return timeElapsed > data.ttl && data.ttl > 0;
-};
+}
